Made all existing singletons to be derivatives of the patterns/Singleton
