+++ src/input/CMakeLists.txt
@@ -19,6 +19,8 @@ set_target_properties(Input PROPERTIES
 )
 
 target_link_libraries(Input
-    PRIVATE 
+    PUBLIC
+        Patterns
+    PRIVATE
         $<$<BOOL:${SPELUNKY_PSP_PLATFORM_DESKTOP}>:SDL_1_XX>
         )
