+++ src/level/interface/Level.hpp
@@ -1,19 +1,12 @@
 #pragma once
 
+#include "patterns/Singleton.hpp"
 #include "TileBatch.hpp"
 
-class Level
+class Level : public Singleton<Level>
 {
 public:
-
-    static Level& instance();
-    static void init();
-    static void dispose();
-
     TileBatch& get_tile_batch() { return _tile_batch; };
-
 private:
-
     TileBatch _tile_batch;
-    static Level* _instance;
 };
