--- src/patterns/interface/Singleton.hpp
@@ -1,12 +1,0 @@
-#pragma once
-
-class Singleton
-{
-public:
-    static Singleton& instance();
-    static void init();
-    static void dispose();
-
-private:
-    static Singleton* _instance;
-};
