+++ src/renderer/CMakeLists.txt
@@ -24,6 +24,7 @@ set_target_properties(Renderer PROPERTIES
 
 target_link_libraries(Renderer
     PUBLIC
+        Patterns
         TextureBank
     PRIVATE
         Logger
