+++ src/texture-bank/CMakeLists.txt
@@ -34,6 +34,7 @@ endif()
 target_link_libraries(TextureBank
     PUBLIC
         Renderer
+        Patterns
     PRIVATE
         Logger
         GraphicsUtils
