+++ src/texture-bank/interface/TextureBank.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "patterns/Singleton.hpp"
 #include "TextureType.hpp"
 #include "TextureRegion.hpp"
 
@@ -12,13 +13,9 @@ using TextureID = unsigned int;
 const TextureID INVALID_TEXTURE = 0;
 const TextureRegion INVALID_REGION = {};
 
-class TextureBank
+class TextureBank : public Singleton<TextureBank>
 {
 public:
-
-    static TextureBank &instance();
-    static void init();
-    static void dispose();
 
     void load_textures();
     void load_texture_regions();
@@ -46,7 +43,6 @@ public:
 
 private:
 
-    static TextureBank *_instance;
     std::unordered_map<TextureType, TextureID> _texture_ids;
     std::unordered_map<TextureType, std::vector<TextureRegion>> _texture_regions;
 };
